backend/requirements.txt
fastapi
uvicorn
pdfplumber
langchain
langchain-openai
openai
bcrypt
python-jose[cryptography]
python-multipart
sqlalchemy
python-dotenv
requests
pydantic
pydantic-settings
faiss-cpu
tiktoken
langchain-community
